pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Feras Alrimali <feras.gogo@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
pydantic = "^2.8.0"
sqlalchemy = "^2.0.31"
sqlmodel = "^0.0.19"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
anyio = "^4.4.0"
uvicorn = "^0.30.1"


netmiko = "^4.3.0"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytz = "^2024.1"
requests = "^2.32.3"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:start"